+++ src/main/java/org/elasql/procedure/DdStoredProcedureFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package org.elasql.schedule;
+package org.elasql.procedure;
 
 
 public interface DdStoredProcedureFactory {
