+++ src/main/java/org/elasql/procedure/tpart/TPartStoredProcedure.java
@@ -1,6 +1,6 @@
-package org.elasql.schedule.tpart;
+package org.elasql.procedure.tpart;
 
-import org.elasql.schedule.DdStoredProcedure;
+import org.elasql.procedure.DdStoredProcedure;
 import org.elasql.schedule.tpart.sink.SunkPlan;
 import org.elasql.sql.RecordKey;
